+++ packages/maps/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.1.1
 - Refactor `MapService` to `GlobalMap`
-- Add `leaflet.heat`
+- Add `leaflet.heat` and `heatmapMixin`
+- Add `actionsHeatmapTrace` and `StopsHeatmapTrace`
+- Add `actionsClusterTrace`
 
 # 0.1.0-beta
 - Fix `rxjs` operator bundling for umd bundle